Add routes to show and delete user with a bit of styling

+++ app/routes/users_route.rb
@@ -2,6 +2,7 @@ module Sinatra
   module Web
     module Routing
       module UsersRoute
+
         def self.registered(app)
           app.get '/signup' do
             if logged_in?
@@ -13,19 +14,28 @@ module Sinatra
           end
 
           app.post '/signup' do
-            if params[:password] == params[:confirm_password]
-              user = User.new(name: params[:name], password: params[:password])
-              if user.save
-                session.delete(:failed_name)
-                session[:id] = user.id
-                url_redirect
+            if params[:password].length > 7
+              if params[:password] == params[:confirm_password]
+                if params[:name].match(/\A[A-Za-z0-9]+\z/) && params[:name].length < 15
+                  user = User.new(name: params[:name], password: params[:password])
+                  if user.save
+                    session.delete(:failed_name)
+                    session[:id] = user.id
+                    url_redirect
+                  else
+                    session[:failed_name] = params[:name]
+                    haml :'users/signup', locals: {message: "Registration invalid. Please try again"}
+                  end
+                else
+                  haml :'users/signup', locals: {message: "Registration invalid. Please try again"}
+                end
               else
                 session[:failed_name] = params[:name]
-                haml :'users/signup', locals: {message: "Registration invalid. Please try again"}
+                haml :'users/signup', locals: {message: "Passwords did not match. Please try again"}
               end
             else
               session[:failed_name] = params[:name]
-              haml :'users/signup', locals: {message: "Passwords did not match. Please try again"}
+              haml :'users/signup', locals: {message: "Password must be at least 8 characters long"}
             end
           end
 
@@ -60,9 +70,51 @@ module Sinatra
           end
 
           app.get '/users/:name' do
-            @user = User.find_by(params)
-            erb :'users/show'
+            @user = current_user if logged_in?
+            @view_user = User.find_by(name: params[:name])
+            if @view_user
+              haml :'users/show'
+            else
+              status 404
+            end
           end
+
+          app.delete '/users/:name/delete' do
+            if logged_in?
+              user = current_user
+              delete_user = User.find_by(name: params[:name])
+              if user && delete_user == user
+                user.destroy
+                session.clear
+                redirect '/'
+              else
+                status 400
+              end
+            else
+              session[:redir] = "/users/#{params[:name]}"
+              redirect '/login'
+            end
+          end
+
+          app.delete '/users/:name/delete/all' do
+            if logged_in?
+              user = current_user
+              delete_user = User.find_by(name: params[:name])
+              if user && delete_user == user
+                user.comments.destroy_all
+                user.topics.destroy_all
+                user.destroy
+                session.clear
+                redirect '/'
+              else
+                status 400
+              end
+            else
+              session[:redir] = "/users/#{params[:name]}"
+              redirect '/login'
+            end
+          end
+
         end
       end
     end
